refactor(scripts): move terraform apply-path guards into scripts/terraform/ (P10)
Nine Terraform guards and apply-pipeline notifiers move into the pipeline they
gate. Every contract-pinned file stays flat: tf-stacks, the deploy-staging
contract, the metrics-bridge bootstrap plan, sanitize-terraform-output, and
verify-github-environment-protection keep the exact paths seven
production-infra-identity-contract files and the Sentry suite manifest name.

Six of the nine moves are byte-identical. Three carry only a depth fix:
check-metrics-bridge-template-plan.mjs imports ../lib/hcl.mjs,
terraform-fmt-check.mjs walks up one more level to the repo root (and prints
its new usage path), and check-terraform-deploy-queue.test.mjs reads
../../.github/workflows/terraform-deploy-queue-watch.yml.

Three silent-failure surfaces move with them. implementation_signature() in
agent-quality-gate.sh hashes an explicit path list; a stale entry hashes as
__missing__, freezing the signature so --skip-if-fresh reuses a dead stamp.
is_quality_setup_command()'s literal pattern keeps the fmt check in the
fail-fast prerequisite phase instead of demoting it into the parallel pool.
Eleven enumerated paths-filter lines across six workflows stop firing their
jobs without failing anything. The gate test now derives the setup-command
pattern from the gate source and matches it against the command the gate
actually emits, with a negative control on the pre-move path.

terraform.stacks.json pins these paths too — nine entries across six stacks,
asserted by tf-stacks.test.mjs. That pin was not recorded anywhere; it is now
listed in scripts/AGENTS.md and in the ADR 0064 sweep checklist.

scripts/deploy-staging-contract.test.mjs gains its own pnpm alias so the
largest suite in the tree no longer depends on tf:test's side-effect import
for its only runner.
