+++ .github/workflows/aegis-terraform.yml
@@ -50,7 +50,7 @@ on:
     paths:
       - aegis/terraform/**
       - scripts/sanitize-terraform-output.sh
-      - scripts/notify-terraform-apply.mjs
+      - scripts/terraform/notify-terraform-apply.mjs
       - scripts/verify-github-environment-protection.mjs
       - scripts/verify-github-environment-protection.test.mjs
       - .github/workflows/aegis-terraform.yml
@@ -59,7 +59,7 @@ on:
     paths:
       - aegis/terraform/**
       - scripts/sanitize-terraform-output.sh
-      - scripts/notify-terraform-apply.mjs
+      - scripts/terraform/notify-terraform-apply.mjs
       - scripts/verify-github-environment-protection.mjs
       - scripts/verify-github-environment-protection.test.mjs
       - .github/workflows/aegis-terraform.yml
@@ -272,7 +272,7 @@ jobs:
           TERRAFORM_PLAN_FILE: /tmp/tf-plan.txt
           TERRAFORM_STACK_LABEL: aegis/terraform
           TERRAFORM_TARGET_ENVIRONMENT: production-infra
-        run: node "$GITHUB_WORKSPACE/scripts/notify-terraform-apply.mjs"
+        run: node "$GITHUB_WORKSPACE/scripts/terraform/notify-terraform-apply.mjs"
 
       - name: Post plan as sticky PR comment
         if: always() && github.event_name == 'pull_request'
