+++ .github/workflows/ci.yml
@@ -178,8 +178,8 @@ jobs:
               - aegis/terraform/**
               - governance-watchdog/**
               - terraform.stacks.json
-              - scripts/check-metrics-bridge-template-plan.mjs
-              - scripts/check-metrics-bridge-template-plan.test.mjs
+              - scripts/terraform/check-metrics-bridge-template-plan.mjs
+              - scripts/terraform/check-metrics-bridge-template-plan.test.mjs
               - scripts/check-peg-policy-publication.mjs
               - scripts/check-peg-policy-publication.test.mjs
               - scripts/production-infra-identity-contract/**
@@ -189,8 +189,8 @@ jobs:
               # their own entries or a parser change stops firing these jobs.
               - scripts/lib/hcl.mjs
               - scripts/lib/workflow-yaml.mjs
-              - scripts/terraform-fmt-check.mjs
-              - scripts/tf-platform-plan-guard.mjs
+              - scripts/terraform/terraform-fmt-check.mjs
+              - scripts/terraform/tf-platform-plan-guard.mjs
               - scripts/tf-stacks.mjs
               - scripts/tf-stacks.test.mjs
               - .github/workflows/alerts-infra.yml
@@ -1030,7 +1030,7 @@ jobs:
       - name: Terraform source-format target regression suite
         # Proves ignored operator tfvars stay outside the explicit format target
         # list while tracked and non-ignored source paths remain covered.
-        run: node scripts/terraform-fmt-check.test.mjs
+        run: node scripts/terraform/terraform-fmt-check.test.mjs
       - name: Hermetic Vitest setup consistency
         run: node scripts/check-hermetic-vitest-setup.mjs
       - name: Hermetic Vitest setup consistency suite
