+++ .github/workflows/governance-watchdog.yml
@@ -79,7 +79,7 @@ on:
       - governance-watchdog/tsconfig.json
       - governance-watchdog/tsconfig.build.json
       - scripts/sanitize-terraform-output.sh
-      - scripts/notify-terraform-apply.mjs
+      - scripts/terraform/notify-terraform-apply.mjs
       - scripts/verify-github-environment-protection.mjs
       - scripts/verify-github-environment-protection.test.mjs
       - .github/workflows/governance-watchdog.yml
@@ -94,7 +94,7 @@ on:
       - governance-watchdog/tsconfig.json
       - governance-watchdog/tsconfig.build.json
       - scripts/sanitize-terraform-output.sh
-      - scripts/notify-terraform-apply.mjs
+      - scripts/terraform/notify-terraform-apply.mjs
       - scripts/verify-github-environment-protection.mjs
       - scripts/verify-github-environment-protection.test.mjs
       - .github/workflows/governance-watchdog.yml
@@ -428,7 +428,7 @@ jobs:
           TERRAFORM_PLAN_FILE: /tmp/tf-plan.txt
           TERRAFORM_STACK_LABEL: governance-watchdog/infra
           TERRAFORM_TARGET_ENVIRONMENT: production-infra
-        run: node "$GITHUB_WORKSPACE/scripts/notify-terraform-apply.mjs"
+        run: node "$GITHUB_WORKSPACE/scripts/terraform/notify-terraform-apply.mjs"
 
       - name: Post plan as sticky PR comment
         if: always() && github.event_name == 'pull_request'
