+++ scripts/AGENTS.md
@@ -17,7 +17,7 @@ garden_lane: agent-entry-points
 ## Scope
 
 `scripts/` holds deploy wrappers, agent quality gates, code-health checks, and
-repo maintenance utilities. 183 files sit flat at the top level today.
+repo maintenance utilities. 174 files sit flat at the top level today.
 
 ## Target Layout
 
@@ -40,7 +40,7 @@ Files stay flat until their phase merges.
 | `terraform/`    | P10   | movable Terraform guards and helpers   |
 | `gate/`         | P11   | quality-gate satellites                |
 
-Landed: P1, P2, P6, P7. `lib/` (the shared tier) and
+Landed: P1, P2, P6, P7, P10. `lib/` (the shared tier) and
 `production-infra-identity-contract/` predate the reorganization. `setup.sh`
 stays flat: `.config/wt.toml` runs that exact path as the Worktrunk pre-start
 hook, and eight docs name it.
@@ -56,7 +56,7 @@ pinned hashes, and identities stay with their domain.
 
 ## Why Files Stay Flat
 
-Seven mechanisms pin `scripts/` paths. A file one of them names moves only when
+Eight mechanisms pin `scripts/` paths. A file one of them names moves only when
 that mechanism moves with it, in the same PR.
 
 - **Autoreview runtime materialization.** `agent-autoreview.sh` names its
@@ -79,6 +79,8 @@ that mechanism moves with it, in the same PR.
   there asserts all three. A miss is silent — the job stops running while the
   required `ci` sentinel stays green. ADR 0064 covers when a module glob such as
   `supply-chain.yml`'s `scripts/supply-chain/**` is the safer pin.
+- **Terraform stack registry.** `terraform.stacks.json` pins exact `scripts/`
+  paths per stack.
 - **Production infrastructure contract pins.**
   `production-infra-identity-contract/workflow-inventory.mjs` pins exact script
   paths for the workflows it audits.
@@ -98,7 +100,7 @@ breaks silently on the next move.
 
 ## Sweep Checklist for a Move
 
-Work the eight-surface checklist in
+Work the nine-surface checklist in
 [ADR 0064](../docs/adr/0064-scripts-module-directories.md#sweep-checklist-for-a-move)
 in the PR that moves a file. Every surface there is mandatory.
