+++ scripts/terraform/check-metrics-bridge-template-plan.test.mjs
