+++ scripts/terraform/tf-platform-plan-guard.mjs
