+++ terraform.stacks.json
@@ -29,10 +29,10 @@
       "changedPathPatterns": [
         "terraform/**",
         "terraform.stacks.json",
-        "scripts/check-metrics-bridge-template-plan.mjs",
-        "scripts/check-metrics-bridge-template-plan.test.mjs",
-        "scripts/terraform-fmt-check.mjs",
-        "scripts/tf-platform-plan-guard.mjs",
+        "scripts/terraform/check-metrics-bridge-template-plan.mjs",
+        "scripts/terraform/check-metrics-bridge-template-plan.test.mjs",
+        "scripts/terraform/terraform-fmt-check.mjs",
+        "scripts/terraform/tf-platform-plan-guard.mjs",
         "scripts/tf-stacks.mjs",
         ".github/workflows/ci.yml",
         ".github/workflows/infra.yml"
@@ -67,7 +67,7 @@
         "scripts/check-peg-policy-publication.test.mjs",
         "scripts/check-peg-registry-integrity.mjs",
         "scripts/production-infra-identity-contract/**",
-        "scripts/terraform-fmt-check.mjs",
+        "scripts/terraform/terraform-fmt-check.mjs",
         "scripts/tf-stacks.mjs",
         "scripts/tf-stacks.test.mjs",
         ".github/workflows/ci.yml",
@@ -95,7 +95,7 @@
       "changedPathPatterns": [
         "alerts/rules/**",
         "terraform.stacks.json",
-        "scripts/terraform-fmt-check.mjs",
+        "scripts/terraform/terraform-fmt-check.mjs",
         "scripts/tf-stacks.mjs",
         ".github/workflows/ci.yml",
         ".github/workflows/infra.yml",
@@ -133,7 +133,7 @@
       "changedPathPatterns": [
         "alerts/infra/**",
         "terraform.stacks.json",
-        "scripts/terraform-fmt-check.mjs",
+        "scripts/terraform/terraform-fmt-check.mjs",
         "scripts/tf-stacks.mjs",
         ".github/workflows/ci.yml",
         ".github/workflows/infra.yml",
@@ -161,7 +161,7 @@
       "changedPathPatterns": [
         "aegis/terraform/**",
         "terraform.stacks.json",
-        "scripts/terraform-fmt-check.mjs",
+        "scripts/terraform/terraform-fmt-check.mjs",
         "scripts/tf-stacks.mjs",
         ".github/workflows/ci.yml",
         ".github/workflows/infra.yml",
@@ -206,7 +206,7 @@
         "governance-watchdog/tsconfig.json",
         "governance-watchdog/tsconfig.build.json",
         "terraform.stacks.json",
-        "scripts/terraform-fmt-check.mjs",
+        "scripts/terraform/terraform-fmt-check.mjs",
         "scripts/tf-stacks.mjs",
         ".github/workflows/ci.yml",
         ".github/workflows/infra.yml",
